requirements.txt
groq
httpx
feedparser
python-dotenv
